+++ gui/accounts/forms.py
@@ -1,3 +1,4 @@
+import re
 from logging import getLogger
 from django.conf import settings
 from django import forms
@@ -201,6 +202,12 @@ class RegisterForm(forms.ModelForm):
             if User.objects.filter(Q(email__iexact=email) | Q(username__iexact=email)):
                 raise forms.ValidationError(_('This email address is already in use. Please supply a different email '
                                               'address.'))
+
+        email_clean_regex = re.compile("^[A-Za-z0-9@.+_-]+$")
+
+        if email_clean_regex.match(email) is None:
+            raise forms.ValidationError(_('Sorry. Your email address did not pass the validity test.'))
+
         return email